pyproject.toml
[project]
name = "transport-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
]
